refactor: move testing helpers for BigInt to their own file(s)
