+++ test/c/CMakeLists.txt
@@ -28,7 +28,7 @@ find_library (qiskit qiskit_cext PATHS ${CARGO_LIB_DIR} REQUIRED)
 
 # Convert warnings into errors. Note that MSVC and GCC/clang do not match here, and CI will
 # require all to pass without errors.
-if (MSVC) 
+if (MSVC)
     add_compile_options(/W3 /WX)
 else ()
     add_compile_options(-Wall -Wextra -Werror -Wshadow)
@@ -50,7 +50,7 @@ file (
 # test file.
 create_test_sourcelist (source_files test_driver.c ${discovered_tests})
 
-add_library(common STATIC common.c)
+add_library(common STATIC common.c bigint.c)
 # Actually define the test driver program executable to be built...
 add_executable (test_driver ${source_files})
 # ...include the location of the header file...
