+++ test/c/test_classical_expr.c
@@ -10,10 +10,10 @@
 // copyright notice, and modified files need to carry a notice indicating
 // that they have been altered from the originals.
 
+#include "bigint.h"
 #include "common.h"
 #include <complex.h>
 #include <inttypes.h>
-#include <math.h>
 #include <qiskit.h>
 #include <stdbool.h>
 #include <stddef.h>
@@ -33,24 +33,6 @@ QkValue *inner_test_value_biguint(const uint8_t *, size_t, bool, uint32_t);
 void inned_test_old_style_vars(QkExprNode **);
 void *inner_expr_free(QkExprNode *);
 void *inner_value_free(QkValue *);
-
-/*
- * Print QkBigUint values (debug helper)
- */
-static void biguint_debug_print(const QkBigUint *biguint, const char *name) {
-    if (name == NULL)
-        name = "biguint";
-    printf("%s.num_limbs = %zu\n", name, biguint->num_limbs);
-    for (size_t limb_idx = 0; limb_idx < biguint->num_limbs; limb_idx++) {
-        printf("%s[%zu] = [", name, limb_idx);
-        for (size_t i = 0; i < sizeof(uint64_t); i++) {
-            printf("%02" PRIX8, ((uint8_t *)biguint->limbs)[sizeof(uint64_t) * limb_idx + i]);
-            if (i < sizeof(uint64_t) - 1)
-                printf(" ");
-        }
-        printf("]\n");
-    }
-}
 
 /*
  * Test that expression tree structure is captured correctly via the information structs
@@ -575,18 +557,6 @@ cleanup:
         qk_biguint_clear(&biguint_val);
 
     return result;
-}
-
-int biguint_cmp(const QkBigUint a, const QkBigUint b) {
-    ptrdiff_t limb_count_diff = (ptrdiff_t)a.num_limbs - (ptrdiff_t)b.num_limbs;
-    if (limb_count_diff != 0) {
-        return limb_count_diff;
-    }
-    return memcmp(a.limbs, b.limbs, a.num_limbs * sizeof(uint64_t));
-}
-
-bool biguint_eq(const QkBigUint a, const QkBigUint b) {
-    return biguint_cmp(a, b) == 0;
 }
 
 typedef struct {
